requirements.txt
aiohttp==3.7.4
asgiref==3.7.2
async-timeout==3.0.1
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
blinker==1.7.0
bson==0.5.10
certifi==2019.9.11
cffi==1.16.0
channels==3.0.5
channels-redis==4.1.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
commonmark==0.9.1
constantly==23.10.4
cryptography==41.0.7
daphne==3.0.2
defusedxml==0.8.0rc2
Django==5.0
django-admin-inline-paginator==0.4.0
django-cors-headers==4.3.1
django-extensions==3.2.3
django-redis==5.4.0
django-taggit==5.0.1
djangorestframework==3.14.0
drf-yasg==1.21.7
easy-thumbnails==2.8.5
Flask==3.0.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.3
jprq==2.1.0
Markdown==3.5.1
MarkupSafe==2.1.3
msgpack==1.0.7
multidict==6.0.4
oauthlib==3.2.2
packaging==23.2
pdfkit==1.0.0
Pillow==10.1.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
pyOpenSSL==23.3.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==10.14.0
service-identity==23.1.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
websockets==9.1
Werkzeug==3.0.1
whitenoise==6.6.0
yarl==1.9.4
zope.interface==6.1
